moved trim-wasm into parser
