+++ packages/transformer/index-adapter-module/alias-def.spec.js
@@ -1,4 +1,3 @@
-import trimWasm from '../trim-wasm.js'
 import pipe from '../pipe.js'
 import Parser from '../parser/index.js'
 
@@ -19,7 +18,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(adapterModule.meta.modules[1].meta.path()).toEqual([
@@ -42,7 +41,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(adapterModule.meta.modules[1].meta.path()).toEqual([
@@ -66,7 +65,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(
@@ -88,7 +87,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(
