+++ packages/transformer/index-adapter-module/module-def.spec.js
@@ -1,5 +1,4 @@
 import { toMatchTree } from '../matchers.js'
-import trimWasm from '../trim-wasm.js'
 import pipe from '../pipe.js'
 import Parser from '../parser/index.js'
 
@@ -18,7 +17,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(adapterModule.meta.modules[0]).toMatchTree(['adapter', 'module'])
@@ -32,7 +31,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(adapterModule.meta.symbolIndex.modules.$M).toBe(0)
@@ -53,7 +52,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(adapterModule.meta.modules[0]).toMatchTree(['module'])
@@ -67,7 +66,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(adapterModule.meta.symbolIndex.modules.$M).toBe(0)
@@ -84,7 +83,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(adapterModule.meta.modules[0]).toMatchTree([
@@ -111,7 +110,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(adapterModule.meta.symbolIndex.modules.$M).toBe(0)
@@ -141,7 +140,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(adapterModule.meta.modules[0]).toMatchTree([
@@ -160,7 +159,7 @@ describe('index adapter module', () => {
         `
 
         const parser = Parser()
-        const adapterModule = pipe(parser, trimWasm, ([node]) => node)(wat)
+        const adapterModule = pipe(parser, ([node]) => node)(wat)
         indexAdapterModule(adapterModule)
 
         expect(adapterModule.meta.symbolIndex.modules.$M).toBe(0)
